docs(navigation): add KDoc for navigation components
